requirement.txt
flask==3.0.0
python-whois==0.8.0
requests==2.31.0
flask-sqlalchemy==3.1.1
Werkzeug==3.0.0
nslookup==1.7.0